64bitでビルドしたとき、 CSHABase64BitUnitクラスのFinalize関数で警告が出るのを対処
ローカル変数targetSizeの型がsize_tではなくuint32_tになっていたのが警告の原因である
size_t型にすることによって対処した。

+++ CppHashClasses/HSSHABase.hpp
@@ -57,7 +57,7 @@ namespace Base {
 			//0x80をセットした位置が896ビット目(112バイト目) 
 			//以上であればハッシュブロックを実行する
 			if (this->m_MessageAddPosition >= 112) {
-				uint32_t targetSize = this->m_MessageAddPosition;
+				size_t targetSize = this->m_MessageAddPosition;
 				this->BlockProcess ();
 				memset (this->m_MessageBlock, 0, targetSize);
 			}
